+++ sample-app/app/mailboxes/application_mailbox.rb
@@ -1,0 +1,3 @@
+class ApplicationMailbox < ActionMailbox::Base
+  routing /.*/ => :test
+end