+++ sample-app/app/mailboxes/test_mailbox.rb
@@ -1,0 +1,4 @@
+class TestMailbox < ApplicationMailbox
+  def process
+  end
+end