+++ sample-app/db/migrate/20241106152613_create_dynamo_db_sessions_table.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class CreateDynamoDbSessionsTable < ActiveRecord::Migration[7.2]
+  def up
+    options = Rails.application.config.session_options
+    Aws::SessionStore::DynamoDB::Table.create_table(options)
+  end
+
+  def down
+    options = Rails.application.config.session_options
+    Aws::SessionStore::DynamoDB::Table.delete_table(options)
+  end
+end