+++ sample_app_old/test/mailers/previews/my_mailer_preview.rb
@@ -1,4 +1,0 @@
-# Preview all emails at http://localhost:3000/rails/mailers/my_mailer
-class MyMailerPreview < ActionMailer::Preview
-
-end