+++ spec/aws/rails/middleware/elastic_beanstalk_job.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class ElasticBeanstalkJob < ActiveJob::Base
-  queue_as :default
-
-  def perform(); end
-end