+++ spec/aws/rails/middleware/elastic_beanstalk_periodic_task.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class ElasticBeanstalkPeriodicTask < ActiveJob::Base
-  queue_as :default
-
-  def perform(); end
-end