+++ spec/dummy/app/jobs/elastic_beanstalk_job.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class ElasticBeanstalkJob < ApplicationJob
+  queue_as :default
+
+  def perform(); end
+end