+++ spec/dummy/app/jobs/elastic_beanstalk_periodic_task.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class ElasticBeanstalkPeriodicTask < ApplicationJob
+  queue_as :default
+
+  def perform(); end
+end